requirements.txt
asgiref==3.8.1
Django==5.0.7
django-phonenumber-field==8.0.0
phonenumbers==8.13.42
pillow==10.4.0
sqlparse==0.5.1
